src/aica_backend/requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
python-multipart==0.0.20
pydantic[email]==2.12.3
supabase==2.23.0
firecrawl-py==4.5.0
langchain==1.0.3
langchain-anthropic==1.0.1
langchain-huggingface==1.0.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0
PyPDF2==3.0.1
python-docx==1.2.0
mammoth==1.11.0
aiohttp==3.13.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.9.0
sentence-transformers==5.1.2
faiss-cpu==1.12.0
numpy==2.3.4
slowapi==0.1.9
reportlab==4.4.4
